Support tuple-like objects, support int coordinates
Allow the user to pass an array of tuple-like objects (i.e. objects that can be
passed to tuple()) instead of normal tuples. For example, if the user defines
a class Point, they can call

    vw.simplify([Point(1, 2), Point(3, 4), ...])

Also support points with integer coordinates, by casting the input coordinates
to floats.

Add tests